requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cryptography==40.0.2
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.3
openai==0.27.7
packaging==23.1
pathy==0.10.1
pdfminer.six==20221105
preshed==3.0.8
pycparser==2.21
pydantic==1.10.8
PyPDF2==3.0.1
requests==2.31.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.10
thinc-apple-ops==0.1.3
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.6.2
urllib3==2.0.2
wasabi==1.1.1
yarl==1.9.2
